# specs/openapi/v1.yaml
openapi: 3.0.3
info:
  title: US Vibe API
  version: 1.0.0
paths:
...
  /auth/login:
    post:
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentialsRequest"
      responses:
        "200":
          description: Access token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractError"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractError"
...
  /sessions:
    post:
      summary: Create a Backend Solo simulation session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationSessionCreateRequest"
      responses:
        "201":
          description: Session created (currentGate A)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationSessionResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractError"
...
  /api/sessions/{sessionId}/role-gap:
    get:
      summary: Workspace role gap (human vs AI agents)
      operationId: getWorkspaceRoleGap
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceSessionId"
      responses:
        "200":
          description: ok + data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceJsonEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/sessions/{sessionId}/session-profile:
    patch:
      summary: Update human roles for workspace session
      operationId: patchWorkspaceSessionProfile
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceSessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [humanRoleIds]
              properties:
                humanRoleIds:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: ok + data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceJsonEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/sessions/{sessionId}/workspace-gates:
    get:
      summary: Prompt spec, contract, PR gate flags
      operationId: getWorkspaceGates
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceSessionId"
      responses:
        "200":
          description: ok + data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceJsonEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/sessions/{sessionId}/workspace-dod-verify:
    post:
      summary: Run repo validate-api-contract + workspace gate checks
      operationId: postWorkspaceDodVerify
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceSessionId"
      responses:
        "200":
          description: ok + data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceJsonEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/sessions/{sessionId}/prompt-spec/convert:
    post:
      summary: Prompt to spec template
      operationId: postWorkspacePromptSpecConvert
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceSessionId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                promptText:
                  type: string
      responses:
        "200":
          description: ok + data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceJsonEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/sessions/{sessionId}/prompt-spec/approve:
    post:
      summary: Approve prompt spec (persists gate)
      operationId: postWorkspacePromptSpecApprove
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceSessionId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                specVersion:
                  type: integer
      responses:
        "200":
          description: ok + data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceJsonEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/sessions/{sessionId}/contract/validate:
    post:
      summary: Run mock OpenAPI validation
      operationId: postWorkspaceContractValidate
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceSessionId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                openApiYaml:
                  type: string
      responses:
        "200":
          description: ok + data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceJsonEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/sessions/{sessionId}/contract/approve:
    post:
      summary: Approve contract after validation
      operationId: postWorkspaceContractApprove
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceSessionId"
      responses:
        "200":
          description: ok + data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceJsonEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/sessions/{sessionId}/retro/reports:
    get:
      summary: List retro reports
      operationId: getWorkspaceRetroReports
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceSessionId"
      responses:
        "200":
          description: ok + data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceJsonEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/sessions/{sessionId}/retro/generate:
    post:
      summary: Generate retro report (rule-based KPI)
      operationId: postWorkspaceRetroGenerate
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceSessionId"
      responses:
        "200":
          description: ok + data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceJsonEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  parameters:
    WorkspaceSessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContractError"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    WorkspaceJsonEnvelope:
      type: object
      description: Workspace session endpoints return JSON with ok and optional data/code
      additionalProperties: true
    AuthCredentialsRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    AuthTokenResponse:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
...
    ContractError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    SimulationSessionCreateRequest:
      type: object
      required: [learningGoal, topic, sprintDuration, skillLevel]
      properties:
        learnerRole:
          type: string
        learningGoal:
          type: string
        topic:
          type: string
        sprintDuration:
          type: string
        skillLevel:
          type: string
        activeRoles:
          type: array
          items:
            type: string
...
    SimulationSessionResponse:
      type: object
      required:
        - id
        - learnerRole
        - learningGoal
        - topic
        - sprintDuration
        - skillLevel
        - activeRoles
        - currentGate
        - gateHistory
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        learnerRole:
          type: string
        learningGoal:
          type: string
        topic:
          type: string
        sprintDuration:
          type: string
        skillLevel:
          type: string
        activeRoles:
          type: array
          items:
            type: string
        currentGate:
          type: string
          enum: [A, B, C, D, DONE]
        gateHistory:
          type: array
          items:
            type: object
            required: [at, from, to, reason]
            properties:
              at:
                type: string
                format: date-time
              from:
                type: string
              to:
                type: string
              reason:
                type: string
        implementationNotes:
          type: string
          nullable: true
        implementationAcknowledgedAt:
          type: string
          format: date-time
          nullable: true
        retroSummary:
          type: object
          nullable: true
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...